Move tests from implementations into tests/
