+++ ndhistogram/tests/mod.rs
@@ -16,10 +16,12 @@ mod test_serialization;
 mod test_trait_object_safety;
 mod test_uniform_axis;
 mod test_uniform_axis_integer;
+mod test_uniformcyclic_axis;
 mod test_value_mean;
 mod test_value_sum;
 mod test_value_weightedmean;
 mod test_value_weightedsum;
 mod test_variable_axis;
+mod test_variablecyclic_axis;
 mod test_variablenoflow_axis;
 mod test_versions;
